requirements.txt
Django==2.2.13
djangorestframework
psycopg2
django-rest-auth
